Hooks implementation done
