+++ src/pages/PostList/styles.css
@@ -2,7 +2,6 @@
   display: flex;
   flex-flow: column;
   width: 600px;
-  height: 100vh;
   border-left: 1px solid #2f3336;
   border-right: 1px solid #2f3336;
   border-bottom: 1px solid #2f3336;
