--- src/pages/PostList/utils.ts
@@ -1,8 +1,0 @@
-export const includesInsensitive = (
-  value: string,
-  matchValue: string
-): boolean => {
-  if (!value) return false;
-
-  return value.toLowerCase().includes(matchValue.toLowerCase());
-};
